Restructured packages and removes deleteFile proto.
