--- gcs/deleteFile/proto/deleteFile.proto
@@ -1,54 +1,0 @@
-// Copyright 2020 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// Author: Evan Spendlove (@evanSpendlove)
-//
-// The DeleteFileProbe proto describes the information necessary for a probe to delete a
-// file in GCS.
-
-syntax = "proto2";
-
-package hermes.deletefile;
-
-option go_package = "github.com/googleinterns/step224-2020/gcs/deleteFile/proto";
-
-import "github.com/google/cloudprober/probes/proto/config.proto";
-import "github.com/googleinterns/step224-2020/config/proto/probe.proto";
-
-// DeleteFileProbe describes the information necessary for a probe to delete a
-// file in GCS.
-message DeleteFileProbe {
-
-  // REQUIRED
-  // HermesProbeDef includes extra information, such as additional metrics, that
-  // is not included in the config.proto supplied by Cloudprober.
-  optional hermes.HermesProbeDef probedef = 1;
-
-  // REQUIRED
-  // The name of the bucket that contains the object to be deleted.
-  optional string bucketName = 2;  
-
-  // REQUIRED
-  // The name of the object to be deleted.
-  optional string objectName = 3;
-
-  // Must extend ProbeDef so this probe can be added to Cloudprober as an
-  // extension.
-  extend cloudprober.probes.ProbeDef {
-    // Extension code explanation:
-    // 200-299 = Codes for GCS
-    // 220-229 = Codes for deleting files in GCS
-    optional DeleteFileProbe delete_file_probe = 220;  // 200 because GCS, 20 because Delete probe
-  }
-}
