+++ hermes/shared/file_gen_test.go
@@ -7,11 +7,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-
-// Author: Alicja Kwiecinska (kwiecinskaa@google.com) github: alicjakwie
-
+//
 // Test suite for file_gen.go
-package hermes
+package shared
 
 import (
 	"fmt"
